+++ ContentPatcher/docs/release-notes.md
@@ -9,6 +9,9 @@ When releasing a format change, don't forget to update the smapi.io/json schema!
 
 
 -->
+## Upcoming release
+* Fixed 'no translation' tooltips shown for config fields with no description.
+
 ## 1.24.1
 Released 31 October 2021 for SMAPI 3.12.6 or later.